+++ example/a.ts
@@ -1,6 +1,29 @@
-import { Elysia } from '../src'
+import { Elysia } from 'elysia'
 
-new Elysia().onError(({ code, error }) => {
-	if (code === 'VALIDATION')
-		return error.all[0].message
-})
+const findUser = (authorization?: string) => {
+	return {
+		name: 'Jane Doe',
+		role: 'admin' as const
+	}
+}
+// ---cut---
+
+const app = new Elysia()
+	.macro({
+		role: (role: 'user' | 'admin') => ({
+			resolve({ status, headers: { authorization } }) {
+				const user = findUser(authorization)
+
+				if(user.role !== role)
+					return status(401)
+
+				return {
+					user
+				}
+			}
+		})
+	})
+	.get('/token', ({ user }) => user, {
+	//                 ^?
+		role: 'admin'
+	})